Fe/feature 897/add css

See merge request s11-webmobile2-sub2/S11P12B308!160

+++ src/components/Sns/BtnList.js
@@ -1,6 +1,7 @@
 import React, { useState } from "react";
 import { useNavigate } from "react-router-dom";
 import API from "../../apis/api";
+import useAuthStore from '../../stores/member';
 
 // import bookMarkSelect from "../../assets/icon/bookmark-select.svg";
 // import bookMarkDefault from "../../assets/icon/bookmark-default.svg";
@@ -18,35 +19,40 @@ import './BtnList.css';
 
 const BtnList = ({ likeCnt: initialLikeCnt, isLiked: initialIsLiked, commentCnt, isBookmarked: initialIsBookmarked, articleId }) => {
   const navigate = useNavigate();
+  const { userData } = useAuthStore();
   const [likeCnt, setLikeCnt] = useState(initialLikeCnt);
   const [isLiked, setIsLiked] = useState(initialIsLiked);
   const [isBookmarked, setIsBookmarked] = useState(initialIsBookmarked);
   const imgStyle = { width: '24px', height: '24px' };
 
   const handleLike = async () => {
-    try {
-      const response = isLiked
-        ? await API.delete(`/user/sns/like/${articleId}`)
-        : await API.post(`/user/sns/like/${articleId}`);
-
-      setIsLiked(!isLiked);
-      setLikeCnt(prevLikeCnt => isLiked ? prevLikeCnt - 1 : prevLikeCnt + 1);
-      console.log('좋아요 성공:', response.data);
-    } catch (error) {
-      console.error('좋아요 실패:', error.response);
+    if (userData) {
+      try {
+        const response = isLiked
+          ? await API.delete(`/user/sns/like/${articleId}`)
+          : await API.post(`/user/sns/like/${articleId}`);
+  
+        setIsLiked(!isLiked);
+        setLikeCnt(prevLikeCnt => isLiked ? prevLikeCnt - 1 : prevLikeCnt + 1);
+        console.log('좋아요 성공:', response.data);
+      } catch (error) {
+        console.error('좋아요 실패:', error.response);
+      }
     }
   };
 
   const handleBookmark = async () => {
-    try {
-      const response = isBookmarked
-        ? await API.delete(`/user/sns/bookmark/${articleId}`)
-        : await API.post(`/user/sns/bookmark/${articleId}`);
-
-      setIsBookmarked(!isBookmarked);
-      console.log('북마크 성공:', response.data);
-    } catch (error) {
-      console.error('북마크 실패:', error.response);
+    if (userData) {
+      try {
+        const response = isBookmarked
+          ? await API.delete(`/user/sns/bookmark/${articleId}`)
+          : await API.post(`/user/sns/bookmark/${articleId}`);
+  
+        setIsBookmarked(!isBookmarked);
+        console.log('북마크 성공:', response.data);
+      } catch (error) {
+        console.error('북마크 실패:', error.response);
+      }
     }
   };
